requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymongo
httpx
cachetools 